Introducing the new MatrixArray object.
The MatrixArray is a construct that conceptually holds and array of
matricies. However it does so by have arrays for each different comonent
of the matricies. In effect there is one array of a11 elemnts, a12
elements, so on a so forth. When operations are performed it generates
MatrixArrayResult objects, which hold the operands of all operations to
be performed. When opAssign is called on a MatrixArray and a
MatrixArrayResult is passed in, the operations are layed out inline with
mixin. This allows for an increadibly high degree of SIMD vectorization.

More work on this to come.