Removed some commented out code and made changes to the diffuseFlux test
Changes to the diffuseFlux test case are intended to simulate the affect
of stacking up a large number of operations that depend on less data
to determine the effect on required memory bandwidth. With these changes
the MatrixArray vastly out performs the old array of matricies. This
still does not contain a representitive case of random element access
that are prevalent in CFD applications however, so it still may not be
a good test. More to come.